GPECN1DTDPBC version 0.5.3
Converted README to README.md. Reset default simulation of the
gpecn1Dtdpbc.sh run script to be one of the example simulations that
exhibits 'An Initial Angular Momentum Dependence of the Sagnac Effect
for Matter Waves'. Also updated CHANGELOG and removed legacy TESTED
sections in the header of the Makefile.